decentralize app logic to context providers
